feat: Multi-platform codegen fixes, silent error warnings, .a80 pass-through
Agon Light 2 / CP/M / ZX Spectrum codegen improvements:
- Fix eZ80 address widening (uint16->24-bit for ADL mode)
- Fix @extern syntax in mos.minz (use extern fun...at addr)
- Sanitize all SMC/TSMC labels ($, ., - -> underscore)
- Fix storeFromHL() for single-byte registers (A-L + shadow)
- Fix MIR trackConstants() for register reuse across scopes
- Fix sanitizeComment() for $ in assembly comments
- Fix Unicode arrows in loop reroll/peephole comments

New features:
- .a80/.asm/.z80 assembly pass-through in mz CLI
- Silent error warnings for dropped @decorators and unknown attributes
- CP/M auto-load address in mze (--target cpm)
- 3 new CP/M examples (hello, fibonacci, demo)

Documentation:
- ADR-0006: eZ80 ADL address widening
- ADR-0007: Platform-specific newline handling
- Agon/CP/M wins roadmap with priority matrix
- Progress report with test matrix

+++ minzc/pkg/z80asm/assembler_test.go
@@ -292,7 +292,7 @@ func TestSymbols(t *testing.T) {
 	}
 	
 	// Check symbols
-	expectedSymbols := map[string]uint16{
+	expectedSymbols := map[string]int{
 		"START": 0x8000,
 		"VALUE": 42,
 		"LOOP":  0x8006,
